fix: type moved file, move types to own file
