+++ services/booking-api/test/helpers/booking.test.ts
@@ -2,7 +2,7 @@
 const { InternalServerError } = require('@helsingborg-stad/npm-api-error-handling/src/errors');
 
 import booking from '../../src/helpers/booking';
-import { BookingBody } from '../../src/helpers/getCreateBookingBody';
+import { BookingBody } from '../../src/helpers/types';
 import params from '../../src/libs/params';
 import * as request from '../../src/libs/request';
